+++ src/main/java/pl/droidsonroids/gif/GifError.java
@@ -13,84 +13,87 @@ import java.util.Locale;
 @SuppressWarnings("MagicNumber") //error code constants matching native ones
 public enum GifError {
     /**
-     * Special value indicating lack of errors
+     * Special value indicating lack of errors.
      */
     NO_ERROR(0, "No error"),
     /**
-     * Failed to open given input
+     * Failed to open given input.
      */
     OPEN_FAILED(101, "Failed to open given input"),
     /**
-     * Failed to read from given input
+     * Failed to read from given input.
      */
     READ_FAILED(102, "Failed to read from given input"),
     /**
-     * Data is not in GIF format
+     * Data is not in GIF format.
      */
     NOT_GIF_FILE(103, "Data is not in GIF format"),
     /**
-     * No screen descriptor detected
+     * No screen descriptor detected.
      */
     NO_SCRN_DSCR(104, "No screen descriptor detected"),
     /**
-     * No image descriptor detected
+     * No image descriptor detected.
      */
     NO_IMAG_DSCR(105, "No image descriptor detected"),
     /**
-     * Neither global nor local color map found
+     * Neither global nor local color map found.
      */
     NO_COLOR_MAP(106, "Neither global nor local color map found"),
     /**
-     * Wrong record type detected
+     * Wrong record type detected.
      */
     WRONG_RECORD(107, "Wrong record type detected"),
     /**
-     * Number of pixels bigger than width * height
+     * Number of pixels bigger than width * height.
      */
     DATA_TOO_BIG(108, "Number of pixels bigger than width * height"),
     /**
-     * Failed to allocate required memory
+     * Failed to allocate required memory.
      */
     NOT_ENOUGH_MEM(109, "Failed to allocate required memory"),
     /**
-     * Failed to close given input
+     * Failed to close given input.
      */
     CLOSE_FAILED(110, "Failed to close given input"),
     /**
-     * Given file was not opened for read
+     * Given file was not opened for read.
      */
     NOT_READABLE(111, "Given file was not opened for read"),
     /**
-     * Image is defective, decoding aborted
+     * Image is defective, decoding aborted.
      */
     IMAGE_DEFECT(112, "Image is defective, decoding aborted"),
     /**
      * Image EOF detected before image complete.
-     * EOF means GIF terminator, NOT the end of input source.
+     * EOF means GIF terminator, not the end of input source.
      */
     EOF_TOO_SOON(113, "Image EOF detected before image complete"),
     /**
-     * No frames found, at least one frame required
+     * No frames found, at least one frame required.
      */
     NO_FRAMES(1000, "No frames found, at least one frame required"),
     /**
-     * Invalid screen size, dimensions must be positive
+     * Invalid screen size, dimensions must be positive.
      */
     INVALID_SCR_DIMS(1001, "Invalid screen size, dimensions must be positive"),
     /**
-     * Invalid image size, dimensions must be positive
+     * Invalid image size, dimensions must be positive.
+     * @deprecated This error is no longer thrown.
      */
+    @Deprecated
     INVALID_IMG_DIMS(1002, "Invalid image size, dimensions must be positive"),
     /**
-     * Image size exceeds screen size
+     * Image size exceeds screen size. Occurs only if input source changes after opening.
+     * Otherwise canvas is extended.
      */
     IMG_NOT_CONFINED(1003, "Image size exceeds screen size"),
     /**
-     * Input source rewind has failed, animation is stopped
+     * Input source rewind has failed, animation is stopped.
      */
     REWIND_FAILED(1004, "Input source rewind has failed, animation is stopped"),
     /**
-     * Invalid and/or indirect byte buffer specified
+     * Invalid and/or indirect byte buffer specified.
      */
     INVALID_BYTE_BUFFER(1005, "Invalid and/or indirect byte buffer specified"),
     /**
@@ -98,7 +101,7 @@ public enum GifError {
      */
     UNKNOWN(-1, "Unknown error");
     /**
-     * Human readable description of the error
+     * Human readable description of the error.
      */
     @NonNull
     public final String description;