+++ src/main/jni/gif.c
@@ -1,6 +1,7 @@
 #include "gif.h"
 
 static ColorMapObject *defaultCmap;
+static const jlong NULL_GIF_INFO = (jlong) (intptr_t) NULL;
 
 /**
 * Global VM reference, initialized in JNI_OnLoad
@@ -103,23 +104,23 @@ static int directByteBufferRewindFun(GifInfo *info) {
     return 0;
 }
 
-__unused JNIEXPORT jobject JNICALL
+__unused JNIEXPORT jlong JNICALL
 Java_pl_droidsonroids_gif_GifInfoHandle_openFile(JNIEnv *env, jclass __unused class,
                                                  jstring jfname, jboolean justDecodeMetaData) {
     if (isSourceNull(jfname, env)) {
-        return NULL;
+        return NULL_GIF_INFO;
     }
 
     const char *const filename = (*env)->GetStringUTFChars(env, jfname, NULL);
     if (filename == NULL) {
         throwException(env, RUNTIME_EXCEPTION_BARE, "GetStringUTFChars failed");
-        return NULL;
+        return NULL_GIF_INFO;
     }
     FILE *file = fopen(filename, "rb");
     (*env)->ReleaseStringUTFChars(env, jfname, filename);
     if (file == NULL) {
         throwGifIOException(D_GIF_ERR_OPEN_FAILED, env);
-        return NULL;
+        return NULL_GIF_INFO;
     }
     GifSourceDescriptor descriptor;
     descriptor.GifFileIn = DGifOpen(file, &fileRead, &descriptor.Error);
@@ -127,29 +128,29 @@ Java_pl_droidsonroids_gif_GifInfoHandle_openFile(JNIEnv *env, jclass __unused cl
     descriptor.startPos = ftell(file);
     struct stat st;
     descriptor.sourceLength = stat(filename, &st) == 0 ? st.st_size : -1;
-    jobject gifHandle = createGifHandle(&descriptor, env, justDecodeMetaData);
-    if (gifHandle == NULL) {
+    GifInfo *info = createGifHandle(&descriptor, env, justDecodeMetaData);
+    if (info == NULL) {
         fclose(file);
     }
-    return gifHandle;
+    return (jlong) (intptr_t) info;
 }
 
-__unused JNIEXPORT jobject JNICALL
+__unused JNIEXPORT jlong JNICALL
 Java_pl_droidsonroids_gif_GifInfoHandle_openByteArray(JNIEnv *env, jclass __unused class,
                                                       jbyteArray bytes, jboolean justDecodeMetaData) {
     if (isSourceNull(bytes, env)) {
-        return NULL;
+        return NULL_GIF_INFO;
     }
     ByteArrayContainer *container = malloc(sizeof(ByteArrayContainer));
     if (container == NULL) {
         throwException(env, OUT_OF_MEMORY_ERROR, OOME_MESSAGE);
-        return NULL;
+        return NULL_GIF_INFO;
     }
     container->buffer = (*env)->NewGlobalRef(env, bytes);
     if (container->buffer == NULL) {
         free(container);
         throwException(env, RUNTIME_EXCEPTION_BARE, "NewGlobalRef failed");
-        return NULL;
+        return NULL_GIF_INFO;
     }
     container->arrLen = (*env)->GetArrayLength(env, container->buffer);
     container->pos = 0;
@@ -159,30 +160,28 @@ Java_pl_droidsonroids_gif_GifInfoHandle_openByteArray(JNIEnv *env, jclass __unus
     descriptor.startPos = container->pos;
     descriptor.sourceLength = container->arrLen;
 
-    jobject gifInfoHandle = createGifHandle(&descriptor, env, justDecodeMetaData);
+    GifInfo *info = createGifHandle(&descriptor, env, justDecodeMetaData);
 
-    if (gifInfoHandle == NULL) {
+    if (info == NULL) {
         (*env)->DeleteGlobalRef(env, container->buffer);
         free(container);
     }
-    return gifInfoHandle;
+    return (jlong) (intptr_t) info;
 }
 
-__unused JNIEXPORT jobject JNICALL
-Java_pl_droidsonroids_gif_GifInfoHandle_openDirectByteBuffer(JNIEnv *env,
-                                                             jclass __unused class, jobject buffer,
-                                                             jboolean justDecodeMetaData) {
+__unused JNIEXPORT jlong JNICALL
+Java_pl_droidsonroids_gif_GifInfoHandle_openDirectByteBuffer(JNIEnv *env, jclass __unused class, jobject buffer, jboolean justDecodeMetaData) {
     jbyte *bytes = (*env)->GetDirectBufferAddress(env, buffer);
     jlong capacity = (*env)->GetDirectBufferCapacity(env, buffer);
     if (bytes == NULL || capacity <= 0) {
         if (!isSourceNull(buffer, env))
             throwGifIOException(D_GIF_ERR_INVALID_BYTE_BUFFER, env);
-        return NULL;
+        return NULL_GIF_INFO;
     }
     DirectByteBufferContainer *container = malloc(sizeof(DirectByteBufferContainer));
     if (container == NULL) {
         throwException(env, OUT_OF_MEMORY_ERROR, OOME_MESSAGE);
-        return NULL;
+        return NULL_GIF_INFO;
     }
     container->bytes = bytes;
     container->capacity = capacity;
@@ -194,21 +193,19 @@ Java_pl_droidsonroids_gif_GifInfoHandle_openDirectByteBuffer(JNIEnv *env,
     descriptor.startPos = container->pos;
     descriptor.sourceLength = container->capacity;
 
-    jobject gifInfoHandle = createGifHandle(&descriptor, env, justDecodeMetaData);
-
-    if (gifInfoHandle == NULL) {
+    GifInfo *info = createGifHandle(&descriptor, env, justDecodeMetaData);
+    if (info == NULL) {
         free(container);
     }
-    return gifInfoHandle;
+    return (jlong) (intptr_t) info;
 }
 
-__unused JNIEXPORT jobject JNICALL
-Java_pl_droidsonroids_gif_GifInfoHandle_openStream(JNIEnv *env, jclass __unused class,
-                                                   jobject stream, jboolean justDecodeMetaData) {
+__unused JNIEXPORT jlong JNICALL
+Java_pl_droidsonroids_gif_GifInfoHandle_openStream(JNIEnv *env, jclass __unused class, jobject stream, jboolean justDecodeMetaData) {
     jclass streamCls = (*env)->NewGlobalRef(env, (*env)->GetObjectClass(env, stream));
     if (streamCls == NULL) {
         throwException(env, RUNTIME_EXCEPTION_BARE, "NewGlobalRef failed");
-        return NULL;
+        return NULL_GIF_INFO;
     }
     jmethodID markMID = (*env)->GetMethodID(env, streamCls, "mark", "(I)V");
     jmethodID readMID = (*env)->GetMethodID(env, streamCls, "read", "([BII)I");
@@ -216,26 +213,26 @@ Java_pl_droidsonroids_gif_GifInfoHandle_openStream(JNIEnv *env, jclass __unused 
 
     if (markMID == 0 || readMID == 0 || resetMID == 0) {
         (*env)->DeleteGlobalRef(env, streamCls);
-        return NULL;
+        return NULL_GIF_INFO;
     }
 
     StreamContainer *container = malloc(sizeof(StreamContainer));
     if (container == NULL) {
         (*env)->DeleteGlobalRef(env, streamCls);
         throwException(env, OUT_OF_MEMORY_ERROR, OOME_MESSAGE);
-        return NULL;
+        return NULL_GIF_INFO;
     }
 
     container->buffer = (*env)->NewByteArray(env, 256);
     if (container->buffer == NULL) {
         (*env)->DeleteGlobalRef(env, streamCls);
         throwException(env, OUT_OF_MEMORY_ERROR, OOME_MESSAGE);
-        return NULL;
+        return NULL_GIF_INFO;
     }
     container->buffer = (*env)->NewGlobalRef(env, container->buffer);
     if (container->buffer == NULL) {
         throwException(env, RUNTIME_EXCEPTION_BARE, "NewGlobalRef failed");
-        return NULL;
+        return NULL_GIF_INFO;
     }
 
     container->readMID = readMID;
@@ -245,7 +242,7 @@ Java_pl_droidsonroids_gif_GifInfoHandle_openStream(JNIEnv *env, jclass __unused 
         free(container);
         (*env)->DeleteGlobalRef(env, streamCls);
         throwException(env, RUNTIME_EXCEPTION_BARE, "NewGlobalRef failed");
-        return NULL;
+        return NULL_GIF_INFO;
     }
     container->streamCls = streamCls;
 
@@ -255,45 +252,40 @@ Java_pl_droidsonroids_gif_GifInfoHandle_openStream(JNIEnv *env, jclass __unused 
     descriptor.rewindFunc = streamRewind;
     descriptor.sourceLength = -1;
 
-    jobject gifInfoHandle;
     (*env)->CallVoidMethod(env, stream, markMID, LONG_MAX);
-    if (!(*env)->ExceptionCheck(env))
-        gifInfoHandle = createGifHandle(&descriptor, env, justDecodeMetaData);
-    else {
-        gifInfoHandle = NULL;
-    }
-
-    if (gifInfoHandle == NULL) {
+    if (!(*env)->ExceptionCheck(env)) {
+        GifInfo *info = createGifHandle(&descriptor, env, justDecodeMetaData);
+        return (jlong) (intptr_t) info;
+    } else {
         (*env)->DeleteGlobalRef(env, streamCls);
         (*env)->DeleteGlobalRef(env, container->stream);
         (*env)->DeleteGlobalRef(env, container->buffer);
         free(container);
+        return NULL_GIF_INFO;
     }
-    return gifInfoHandle;
 }
 
-__unused JNIEXPORT jobject JNICALL
-Java_pl_droidsonroids_gif_GifInfoHandle_openFd(JNIEnv *env, jclass __unused handleClass,
-                                               jobject jfd, jlong offset, jboolean justDecodeMetaData) {
+__unused JNIEXPORT jlong JNICALL
+Java_pl_droidsonroids_gif_GifInfoHandle_openFd(JNIEnv *env, jclass __unused handleClass, jobject jfd, jlong offset, jboolean justDecodeMetaData) {
     if (isSourceNull(jfd, env)) {
-        return NULL;
+        return NULL_GIF_INFO;
     }
     jclass fdClass = (*env)->GetObjectClass(env, jfd);
     jfieldID fdClassDescriptorFieldID = (*env)->GetFieldID(env, fdClass, "descriptor", "I");
     if (fdClassDescriptorFieldID == NULL) {
-        return NULL;
+        return NULL_GIF_INFO;
     }
     const int fd = dup((*env)->GetIntField(env, jfd, fdClassDescriptorFieldID));
     if (fd == -1) {
         throwGifIOException(D_GIF_ERR_OPEN_FAILED, env);
-        return NULL;
+        return NULL_GIF_INFO;
     }
     FILE *file = fdopen(fd, "rb");
     if (file == NULL) {
         throwGifIOException(D_GIF_ERR_OPEN_FAILED, env);
-        return NULL;
+        return NULL_GIF_INFO;
     }
-    jobject gifHandle;
+    GifInfo *info;
     if (fseek(file, offset, SEEK_SET) == 0) {
         GifSourceDescriptor descriptor;
         descriptor.GifFileIn = DGifOpen(file, &fileRead, &descriptor.Error);
@@ -301,15 +293,15 @@ Java_pl_droidsonroids_gif_GifInfoHandle_openFd(JNIEnv *env, jclass __unused hand
         descriptor.startPos = ftell(file);
         struct stat st;
         descriptor.sourceLength = fstat(fd, &st) == 0 ? st.st_size : -1;
-        gifHandle = createGifHandle(&descriptor, env, justDecodeMetaData);
+        info = createGifHandle(&descriptor, env, justDecodeMetaData);
     }
     else {
         fclose(file);
-        gifHandle = NULL;
+        info = NULL;
         throwGifIOException(D_GIF_ERR_OPEN_FAILED, env);
     }
 
-    return gifHandle;
+    return (jlong) (intptr_t) info;
 }
 
 __unused JNIEXPORT void JNICALL
@@ -350,6 +342,57 @@ Java_pl_droidsonroids_gif_GifInfoHandle_free(JNIEnv *env, jclass __unused handle
     info->gifFilePtr->UserData = NULL;
     releaseSurfaceDescriptor(info->surfaceDescriptor, env);
     cleanUp(info);
+}
+
+__unused JNIEXPORT void JNICALL
+Java_pl_droidsonroids_gif_GifInfoHandle_setSampleSize(__unused JNIEnv *env, jclass __unused class, jlong gifInfo, jint sampleSize) {
+    GifInfo *info = (GifInfo *) (intptr_t) gifInfo;
+    if (info == NULL)
+        return;
+
+    info->sampleSize = (uint_fast16_t) sampleSize;
+    info->gifFilePtr->SHeight /= info->sampleSize;
+    info->gifFilePtr->SWidth /= info->sampleSize;
+    if (info->gifFilePtr->SHeight == 0) {
+        info->gifFilePtr->SHeight = 1;
+    }
+    if (info->gifFilePtr->SWidth == 0) {
+        info->gifFilePtr->SWidth = 1;
+    }
+
+    SavedImage *sp;
+    uint_fast32_t i;
+    for (i = 0; i < info->gifFilePtr->ImageCount; i++) {
+        sp = &info->gifFilePtr->SavedImages[i];
+        sp->ImageDesc.Width /= info->sampleSize;
+        sp->ImageDesc.Height /= info->sampleSize;
+        sp->ImageDesc.Left /= info->sampleSize;
+        sp->ImageDesc.Top /= info->sampleSize;
+    }
+}
+
+__unused JNIEXPORT jint JNICALL
+Java_pl_droidsonroids_gif_GifInfoHandle_getWidth(__unused JNIEnv *env, jclass __unused class, jlong gifInfo) {
+    GifInfo *info = (GifInfo *) (intptr_t) gifInfo;
+    if (info == NULL)
+        return 0;
+    return (jint) info->gifFilePtr->SWidth;
+}
+
+__unused JNIEXPORT jint JNICALL
+Java_pl_droidsonroids_gif_GifInfoHandle_getHeight(__unused JNIEnv *env, jclass __unused class, jlong gifInfo) {
+    GifInfo *info = (GifInfo *) (intptr_t) gifInfo;
+    if (info == NULL)
+        return 0;
+    return (jint) info->gifFilePtr->SHeight;
+}
+
+__unused JNIEXPORT jint JNICALL
+Java_pl_droidsonroids_gif_GifInfoHandle_getNumberOfFrames(__unused JNIEnv *env, jclass __unused class, jlong gifInfo) {
+    GifInfo *info = (GifInfo *) (intptr_t) gifInfo;
+    if (info == NULL)
+        return 0;
+    return (jint) info->gifFilePtr->ImageCount;
 }
 
 __unused JNIEXPORT jint JNICALL